Move attribute checker functions in common.
